+++ Assets/Enemies/GeneralStuff/DamageNumbers.prefab
@@ -86,7 +86,7 @@ MonoBehaviour:
   m_FallbackScreenDPI: 96
   m_DefaultSpriteDPI: 96
   m_DynamicPixelsPerUnit: 1
-  m_PresetInfoIsWorld: 1
+  m_PresetInfoIsWorld: 0
 --- !u!114 &2692249247841875697
 MonoBehaviour:
   m_ObjectHideFlags: 0
