+++ Assets/Enemies/TimeBomb/TimeBomb.prefab
@@ -5867,6 +5867,9 @@ PrefabInstance:
     - targetCorrespondingSourceObject: {fileID: 919132149155446097, guid: b69ab7dc4a09f094885bf950a65c44c3, type: 3}
       insertIndex: -1
       addedObject: {fileID: -4401849361817302130}
+    - targetCorrespondingSourceObject: {fileID: 919132149155446097, guid: b69ab7dc4a09f094885bf950a65c44c3, type: 3}
+      insertIndex: -1
+      addedObject: {fileID: -980023486028482642}
   m_SourcePrefab: {fileID: 100100000, guid: b69ab7dc4a09f094885bf950a65c44c3, type: 3}
 --- !u!4 &4703448632437160086 stripped
 Transform:
@@ -5953,9 +5956,9 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 06a1a1f7248159247a0ffa7ba240b402, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  healthBarScript: {fileID: 4892327625391498504}
   dmgCanvas: {fileID: 2583985077653713364, guid: 190a5b9014b023b4eacb14b127fcc0ed, type: 3}
   showDamage: 0
+  canvasScript: {fileID: 0}
 --- !u!114 &-1368203466199457740
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -5975,6 +5978,8 @@ MonoBehaviour:
   clipSize: 0
   reloadTime: 0
   baseDamage: 0
+  baseKB: 0
+  critChance: 0
   runSpeed: 45
   jumpForce: 120
   crouchSpeed: 0
@@ -5994,6 +5999,8 @@ MonoBehaviour:
   dashSpeed: 0
   postDashSpeedReduction: 0
   minPostDashSpeed: 0
+  healthBarScript: {fileID: 0}
+  damageNumbers: {fileID: 0}
 --- !u!114 &-4401849361817302130
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -6010,3 +6017,21 @@ MonoBehaviour:
   creditCost: 15
   minStage: 0
   isFlyer: 0
+--- !u!114 &-980023486028482642
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 5336151141367865900}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 904a2cca710ecdf49832605cc50b1f07, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  allItems: []
+  statBoosts: []
+  onHitEffects: []
+  onKillEffects: []
+  onAbilityEffects: []
+  onHPEffects: []
